// client/src/types/Placement.ts
enum Placement {
  TOP_LEFT = "topLeft",
  TOP_CENTER = "topCenter",
  TOP_RIGHT = "topRight",
  LEFT_TOP = "leftTop",
  LEFT_CENTER = "leftCenter",
  LEFT_BOTTOM = "leftBottom",
  RIGHT_TOP = "rightTop",
  RIGHT_CENTER = "rightCenter",
  RIGHT_BOTTOM = "rightBottom",
  BOTTOM_LEFT = "bottomLeft",
  BOTTOM_CENTER = "bottomCenter",
  BOTTOM_RIGHT = "bottomRight",
}

export default Placement;